restructure fx providers
