+++ go/cmd/controllermgr/main.go
@@ -1,20 +1,14 @@
 package main
 
 import (
-	"fmt"
-
 	"github.com/go-logr/zapr"
 	"go.uber.org/fx"
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/dynamic"
 	kubescheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/tools/record"
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	"github.com/uber-go/tally"
-
-	kubeclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	apiHandler "github.com/michelangelo-ai/michelangelo/go/api/handler"
 	baseconfig "github.com/michelangelo-ai/michelangelo/go/base/config"
@@ -72,43 +66,6 @@ func getTallyScope() (tally.Scope, error) {
 	return s, nil
 }
 
-// newProxyProvider creates a new proxy provider
-func newProxyProvider(dynamicClient dynamic.Interface, logger *zap.Logger) proxy.ProxyProvider {
-	return proxy.NewHTTPRouteManager(dynamicClient, logger)
-}
-
-// newDynamicClient creates a Kubernetes dynamic client for working with unstructured resources
-func newDynamicClient() (dynamic.Interface, error) {
-	restConfig, err := ctrl.GetConfig()
-	if err != nil {
-		panic(fmt.Errorf("failed to get rest config: %w", err))
-	}
-	dynamicClient, err := dynamic.NewForConfig(restConfig)
-	if err != nil {
-		panic(fmt.Errorf("failed to create dynamic client: %w", err))
-	}
-	return dynamicClient, nil
-}
-
-// newModelConfigMapProvider creates a new model config map provider
-func newModelConfigMapProvider(client kubeclient.Client, logger *zap.Logger) configmap.ModelConfigMapProvider {
-	return configmap.NewDefaultModelConfigMapProvider(client, logger)
-}
-
-// newEventRecorder creates a new event recorder
-func newEventRecorder(mgr ctrl.Manager) record.EventRecorder {
-	return mgr.GetEventRecorderFor(inferenceserver.ControllerName)
-}
-
-// newInferenceServerGateway creates a new inference server gateway with clients
-func newInferenceServerGateway(kubeClient kubeclient.Client, modelConfigMapProvider configmap.ModelConfigMapProvider, logger *zap.Logger) gateways.Gateway {
-	return gateways.NewGatewayWithClients(gateways.Params{
-		Logger:                 logger,
-		KubeClient:             kubeClient,
-		ModelConfigMapProvider: modelConfigMapProvider,
-	})
-}
-
 // options provides the FX modules and configurations used by the application.
 //
 // This function defines the dependencies and lifecycle management for the application by:
@@ -123,16 +80,14 @@ func options() fx.Option {
 		env.Module,
 		zapfx.Module,
 		baseconfig.Module,
+		gateways.Module,
+		configmap.Module,
+		proxy.Module,
 		fx.Provide(scheme),
 		fx.Provide(baseconfig.GetK8sConfig),
 		fx.Provide(baseconfig.GetMetadataStorageConfig),
 		fx.Provide(baseconfig.GetWorkflowClientConfig),
 		fx.Provide(getTallyScope),
-		fx.Provide(newProxyProvider),
-		fx.Provide(newDynamicClient),
-		fx.Provide(newModelConfigMapProvider),
-		fx.Provide(newInferenceServerGateway),
-		fx.Provide(newEventRecorder),
 		apiHandler.CtrlMgrModule,
 		spark.Module,
 		ray.Module,
