+++ go/components/inferenceserver/configmap/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "interface.go",
+        "module.go",
         "provider.go",
     ],
     importpath = "github.com/michelangelo-ai/michelangelo/go/components/inferenceserver/configmap",
@@ -12,6 +13,7 @@ go_library(
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
+        "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_zap//:go_default_library",
     ],
 )
