+++ go/components/inferenceserver/gateways/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "gateway.go",
         "interface.go",
+        "module.go",
         "registry.go",
     ],
     importpath = "github.com/michelangelo-ai/michelangelo/go/components/inferenceserver/gateways",
@@ -14,6 +15,7 @@ go_library(
         "//go/components/inferenceserver/configmap:go_default_library",
         "//proto/api/v2:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
+        "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_zap//:go_default_library",
     ],
 )
