+++ go/components/inferenceserver/gateways/module.go
@@ -1,0 +1,22 @@
+package gateways
+
+import (
+	"go.uber.org/fx"
+	"go.uber.org/zap"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/michelangelo-ai/michelangelo/go/components/inferenceserver/configmap"
+)
+
+var Module = fx.Options(
+	fx.Provide(newInferenceServerGateway),
+)
+
+// newInferenceServerGateway creates a new inference server gateway with clients
+func newInferenceServerGateway(kubeClient client.Client, modelConfigMapProvider configmap.ModelConfigMapProvider, logger *zap.Logger) Gateway {
+	return NewGatewayWithClients(Params{
+		Logger:                 logger,
+		KubeClient:             kubeClient,
+		ModelConfigMapProvider: modelConfigMapProvider,
+	})
+}
